uploader: Fix a bug while deleting.
Delete file if only it is referred from nowhere.

+++ app/models/Attachment.java
@@ -186,6 +186,8 @@ public class Attachment extends Model {
     @Override
     public void delete() {
         super.delete();
-        new File(uploadDirectory, this.hash).delete();
+        if (!exists(this.hash)) {
+            new File(uploadDirectory, this.hash).delete();
+        }
     }
 }
